refactor(notes.py): simplify and streamline notes management code
- Remove redundant database operations and logging
- Simplify command structure and usage
- Improve command descriptions and help messages
- Use ModerationCogBase for common moderation functionality
- Remove unused imports and clean up code structure

refactor(notes): refactor note commands to use commands.Context instead of discord.Interaction for better compatibility
feat(notes): add aliases for note commands for better user experience
fix(notes): handle case when note is not found in delete, update and view commands to prevent errors
feat(notes): add error handling for MemberNotFound and UserNotFound exceptions in delete, update and view commands
refactor(notes): refactor handle_note_response to include previous content in update command for better tracking of changes

feat(config.py): update permission level descriptions for better clarity
feat(ban.py, cases.py, jail.py, kick.py): add user avatar to embeds for better user identification
feat(jail.py): add setup_jail command to configure jail role and channel permissions
fix(jail.py): add checks for jail role and channel existence before jailing a user

chore(notes.py): remove unused Notes class to clean up codebase
feat(timeout.py, unban.py, unjail.py, untimeout.py, warn.py): add user avatar to embed thumbnail for better user recognition
refactor(unjail.py): improve error handling and add more checks for jail role and channel existence
style(info.py): remove unnecessary logging statements to declutter logs

refactor(snippets.py): remove page parameter from list_snippets function to simplify usage
feat(snippets.py): implement pagination for list_snippets function to improve user experience
fix(snippets.py): improve error handling for snippet deletion to prevent unauthorized deletions
style(error.py): remove unnecessary blank line to improve code readability
style(event.py): improve warning message for harmful commands to enhance user understanding

fix(embeds.py): remove unnecessary rounding precision for latency calculation to simplify code
style(embeds.py): remove unnecessary special characters from latency display for cleaner output
refactor(embeds.py): remove type ignore comments after confirming type compatibility
fix(flags.py): change default reason for KickFlags and TimeoutFlags to MISSING for better error handling
docs(flags.py): update description for TimeoutFlags reason to include examples
