+++ tux/cogs/moderation/ban.py
@@ -108,6 +108,7 @@ class Ban(ModerationCogBase):
                 color=CONST.EMBED_COLORS["CASE"],
                 icon_url=CONST.EMBED_ICONS["ACTIVE_CASE"],
             )
+            embed.set_thumbnail(url=target.avatar)
         else:
             embed = await self.create_embed(
                 ctx,
