+++ tux/cogs/utility/snippets.py
@@ -5,10 +5,12 @@ import discord
 from discord import AllowedMentions
 from discord.ext import commands
 from loguru import logger
+from reactionmenu import ViewButton, ViewMenu
 
 from prisma.models import Snippet
 from tux.database.controllers import DatabaseController
-from tux.utils.embeds import EmbedCreator
+from tux.utils.constants import Constants as CONST
+from tux.utils.embeds import EmbedCreator, create_embed_footer
 
 
 class Snippets(commands.Cog):
@@ -20,19 +22,17 @@ class Snippets(commands.Cog):
     @commands.command(
         name="snippets",
         aliases=["ls"],
-        usage="$snippets <page>",
+        usage="$snippets",
     )
     @commands.guild_only()
-    async def list_snippets(self, ctx: commands.Context[commands.Bot], page: int = 1) -> None:
+    async def list_snippets(self, ctx: commands.Context[commands.Bot]) -> None:
         """
-        List snippets by page.
+        List snippets by pagination.
 
         Parameters
         ----------
         ctx : commands.Context[commands.Bot]
             The context object.
-        page : int, optional
-            The page number, by default 1.
         """
 
         if ctx.guild is None:
@@ -41,11 +41,8 @@ class Snippets(commands.Cog):
 
         snippets: list[Snippet] = await self.db.get_all_snippets_sorted(newestfirst=True)
 
-        # remove snippets that are not in the current server
+        # Remove snippets that are not in the current server
         snippets = [snippet for snippet in snippets if snippet.guild_id == ctx.guild.id]
-
-        # Calculate the number of pages based on the number of snippets
-        pages = 1 if len(snippets) <= 10 else len(snippets) // 10 + 1
 
         # If there are no snippets, send an error message
         if not snippets:
@@ -57,36 +54,45 @@ class Snippets(commands.Cog):
             await ctx.send(embed=embed, delete_after=5)
             return
 
-        # If the page number is invalid, send an error message
-        if page < 1 or page > pages:
-            embed = EmbedCreator.create_error_embed(
-                title="Error",
-                description="Invalid page number.",
-                ctx=ctx,
-            )
-            await ctx.send(embed=embed)
-            return
+        menu = ViewMenu(ctx, menu_type=ViewMenu.TypeEmbed)
 
-        # Get the snippets for the specified page
-        snippets = snippets[(page - 1) * 10 : page * 10]
+        snippets_per_page = 10
+        for i in range(0, len(snippets), snippets_per_page):
+            embed = self._create_snippets_list_embed(ctx, snippets[i : i + snippets_per_page], len(snippets))
+            menu.add_page(embed)
 
-        # Snippets:
-        # `01. snippet_name        | author: author_name`
-        # `02. longer_snippet_name | author: author_name`
+        menu.add_button(ViewButton.back())
+        menu.add_button(ViewButton.next())
+        menu.add_button(ViewButton.end_session())
 
-        embed = EmbedCreator.create_info_embed(
-            title="Snippets",
-            description="\n".join(
-                [
-                    f"`{str(index + 1).zfill(2)}. {snippet.snippet_name.ljust(20)} | author: {self.bot.get_user(snippet.snippet_user_id) or 'Unknown'}`"
-                    for index, snippet in enumerate(snippets)
-                ],
-            ),
-            ctx=ctx,
+        await menu.start()
+
+    def _create_snippets_list_embed(
+        self,
+        ctx: commands.Context[commands.Bot],
+        snippets: list[Snippet],
+        total_snippets: int,
+    ) -> discord.Embed:
+        embed = discord.Embed(
+            title=f"Total Snippets ({total_snippets})",
+            description="",
+            color=CONST.EMBED_COLORS["DEFAULT"],
         )
-        embed.set_footer(text=f"Page {page}/{pages}")
 
-        await ctx.send(embed=embed)
+        if ctx.guild:
+            embed.set_author(name=ctx.guild.name, icon_url=ctx.guild.icon)
+
+        footer_text, footer_icon_url = create_embed_footer(ctx)
+        embed.set_footer(text=footer_text, icon_url=footer_icon_url)
+        embed.timestamp = ctx.message.created_at
+        description = "```\n"
+        for snippet in snippets:
+            author = self.bot.get_user(snippet.snippet_user_id) or "Unknown"
+            description += f"{snippet.snippet_name.ljust(20)} | by: {author}\n"
+        description += "```"
+        embed.description = description
+
+        return embed
 
     @commands.command(
         name="deletesnippet",
@@ -123,12 +129,21 @@ class Snippets(commands.Cog):
 
         # Check if the author of the snippet is the same as the user who wants to delete it and if theres no author don't allow deletion
 
-        # TODO: this was quick and dirty, needs to be refactored
-
         author_id = snippet.snippet_user_id or 0
         if author_id != ctx.author.id:
             conf = await self.config.get_guild_config(ctx.guild.id)
-            user_roles = [role.id for role in ctx.author.roles]  # type: ignore
+            member = ctx.guild.get_member(ctx.author.id)
+
+            if member is None:
+                embed = EmbedCreator.create_error_embed(
+                    title="Error",
+                    description="You can only delete your own snippets.",
+                    ctx=ctx,
+                )
+                await ctx.send(embed=embed)
+                return
+
+            user_roles = [role.id for role in member.roles]
 
             if not conf:
                 embed = EmbedCreator.create_error_embed(
@@ -139,7 +154,12 @@ class Snippets(commands.Cog):
                 await ctx.send(embed=embed)
                 return
 
-            if all(role not in user_roles for role in conf.perm_level_9_role_id):
+            # TODO: Fix snippet deletion permission check later
+            if (conf.perm_level_9_role_id or 0) | (conf.perm_level_8_role_id or 0) | (
+                conf.perm_level_7_role_id or 0
+            ) | (conf.perm_level_6_role_id or 0) | (conf.perm_level_5_role_id or 0) | (
+                conf.perm_level_4_role_id or 0
+            ) | (conf.perm_level_3_role_id or 0) | (conf.perm_level_2_role_id or 0) not in user_roles:
                 embed = EmbedCreator.create_error_embed(
                     title="Error",
                     description="You can only delete your own snippets.",
@@ -190,57 +210,57 @@ class Snippets(commands.Cog):
 
         await ctx.send(text, allowed_mentions=AllowedMentions.none())
 
-    @commands.command(
-        name="snippetinfo",
-        aliases=["si"],
-        usage="$snippetinfo [name]",
-    )
-    @commands.guild_only()
-    async def get_snippet_info(self, ctx: commands.Context[commands.Bot], name: str) -> None:
-        """
-        Get information about a snippet by name.
+    # @commands.command(
+    #     name="snippetinfo",
+    #     aliases=["si"],
+    #     usage="$snippetinfo [name]",
+    # )
+    # @commands.guild_only()
+    # async def get_snippet_info(self, ctx: commands.Context[commands.Bot], name: str) -> None:
+    #     """
+    #     Get information about a snippet by name.
 
-        Parameters
-        ----------
-        ctx : commands.Context[commands.Bot]
-            The context object.
-        name : str
-            The name of the snippet.
-        """
+    #     Parameters
+    #     ----------
+    #     ctx : commands.Context[commands.Bot]
+    #         The context object.
+    #     name : str
+    #         The name of the snippet.
+    #     """
 
-        if ctx.guild is None:
-            await ctx.send("This command cannot be used in direct messages.")
-            return
+    #     if ctx.guild is None:
+    #         await ctx.send("This command cannot be used in direct messages.")
+    #         return
 
-        snippet = await self.db.get_snippet_by_name_and_guild_id(name, ctx.guild.id)
+    #     snippet = await self.db.get_snippet_by_name_and_guild_id(name, ctx.guild.id)
 
-        if snippet is None:
-            embed = EmbedCreator.create_error_embed(
-                title="Error",
-                description="Snippet not found.",
-                ctx=ctx,
-            )
-            await ctx.send(embed=embed, delete_after=5)
-            return
+    #     if snippet is None:
+    #         embed = EmbedCreator.create_error_embed(
+    #             title="Error",
+    #             description="Snippet not found.",
+    #             ctx=ctx,
+    #         )
+    #         await ctx.send(embed=embed, delete_after=5)
+    #         return
 
-        author = self.bot.get_user(snippet.snippet_user_id) or ctx.author
+    #     author = self.bot.get_user(snippet.snippet_user_id) or ctx.author
 
-        embed: discord.Embed = EmbedCreator.custom_footer_embed(
-            title="Snippet Information",
-            content=f"**Name:** {snippet.snippet_name}\n**Author:** {author}\n**Created At:** (set as embed timestamp)\n**Content:** {snippet.snippet_content}",
-            ctx=ctx,
-            latency="N/A",
-            interaction=None,
-            state="DEBUG",
-            user=author,
-        )
+    #     embed: discord.Embed = EmbedCreator.custom_footer_embed(
+    #         title="Snippet Information",
+    #         content=f"**Name:** {snippet.snippet_name}\n**Author:** {author}\n**Created At:** (set as embed timestamp)\n**Content:** {snippet.snippet_content}",
+    #         ctx=ctx,
+    #         latency="N/A",
+    #         interaction=None,
+    #         state="DEFAULT",
+    #         user=author,
+    #     )
 
-        embed.timestamp = snippet.snippet_created_at or datetime.datetime.fromtimestamp(
-            0,
-            datetime.UTC,
-        )
+    #     embed.timestamp = snippet.snippet_created_at or datetime.datetime.fromtimestamp(
+    #         0,
+    #         datetime.UTC,
+    #     )
 
-        await ctx.send(embed=embed)
+    #     await ctx.send(embed=embed)
 
     @commands.command(
         name="createsnippet",
