+++ tux/handlers/error.py
@@ -107,7 +107,6 @@ class ErrorHandler(commands.Cog):
             return
 
         # Check for original error raised and sent to CommandInvokeError
-
         error = getattr(error, "original", error)
 
         # Get the error message and send it to the user
