+++ tux/handlers/event.py
@@ -23,9 +23,10 @@ class EventHandler(commands.Cog):
             return
 
         stripped_content = strip_formatting(message.content)
+
         if is_harmful(stripped_content):
             await message.reply(
-                "Warning: This command is potentially harmful. Please avoid running it unless you are fully aware of its operation. If this was a mistake, please disregard this message.",
+                "Warning: This command is potentially harmful. Please avoid running it unless you're fully aware of it's operation. If this was a mistake, disregard this message.",
             )
 
     @commands.Cog.listener()
